moved tests for model comparison plot to their own module
